Addon(s) installed in SsimLibrary or Session

Lists the addon SyncroSim package(s) associated with a SsimLibrary or Session. These packages can only be used to extend existing SyncroSim base packages; as a result they cannot be used to create new SsimLibraries. For example, stsimsf is an addon for stsim which provides optional additional functionality for the base ST-Sim model. More information on addons can be found in the syncrosim documentation.

addon(ssimObject)

# S4 method for character
addon(ssimObject)

# S4 method for missingOrNULL
addon(ssimObject)

# S4 method for Session
addon(ssimObject)

# S4 method for SsimObject
addon(ssimObject)

Arguments

ssimObject
SsimLibrary or Session object. If NULL (default), session() will be used

Value

A data.frame listing the addon(s) in use by the SsimLibrary or Session to which the object belongs.

Examples

# \donttest{
# Install the base package "stsim"
addPackage("stsim")

#> Package <stsim> is already installed

# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# Retrieve a data.frame of available add-on(s) for the SsimLibrary
addon(myLibrary)

#>              name
#> 1     stsimecodep
#> 2 stsimresolution
#>                                                                                description
#> 1 Calculates ecological departure in ST-Sim using the LANDFIRE Fire Regime Condition Class
#> 2        An ST-Sim add on that enables multiple raster resolutions for spatial simulations
#>   enabled currentVersion minimumVersion
#> 1   FALSE          3.3.0            N/A
#> 2   FALSE          1.0.0            N/A

# }