Installed or available packages

Retrieves the packages installed or available for this version of SyncroSim.

package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for character
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for missingOrNULL
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for Session
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for SsimLibrary
package(ssimObject)

Arguments

ssimObject

Session or SsimLibrary object. If NULL (default), session() will be used

installed

logical or character. TRUE (default) to list installed packages, FALSE to list available packages, and "BASE" to list installed base packages

listTemplates

character. Name of a SyncroSim package. If not NULL (default), then lists all templates available for that package. The package must be installed in the current Session. Ignored if ssimObject is a SsimLibrary object

Value

Returns a data.frame of packages installed or templates available for a specified package.

Examples

# \donttest{
# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# List all installed packages
package(mySession)
#>                        name
#> 1                      Name
#> 2           burnp3basemodel
#> 3                burnP3Plus
#> 4       burnP3PlusCell2Fire
#> 5      burnP3PlusPrometheus
#> 6                 demosales
#> 7                  designer
#> 8                     dgsim
#> 9                       epi
#> 10       epiModelVocVaccine
#> 11               helloworld
#> 12          helloworldConda
#> 13           helloworldOmni
#> 14       helloworldPipeline
#> 15        helloworldSpatial
#> 16   helloworldSpatialConda
#> 17 helloworldSpatialReports
#> 18           helloworldTime
#> 19    helloworldUncertainty
#> 20        landfirevegmodels
#> 21                  netlogo
#> 22                omniscape
#> 23                stconnect
#> 24                    stsim
#> 25          stsimBurnP3Plus
#> 26             stsimcbmcfs3
#> 27          stsimFsjHabitat
#> 28                stsimpvfc
#> 29                  stsimsf
#> 30                    wisdm
#>                                                                                                       description
#> 1                                                                                                     Description
#> 2                                                                                               BurnP3 base model
#> 3                                                                  Burn-P3+ package for burn probability modeling
#> 4                                                          Cell2Fire fire growth model add-on package for BurnP3+
#> 5                                                         Prometheus fire growth model add-on package for BurnP3+
#> 6                                                                               Demo Sales SyncroSim Base Package
#> 7                                                                                      SyncroSim Package Designer
#> 8                                                                  Simulates demographics of wildlife populations
#> 9                                                                         Model of epidemic infections and deaths
#> 10                                                               Integrates VOCs and vaccination in the Epi model
#> 11                                                                  Example demonstrating how to create a package
#> 12                                          Example demonstrating how to create a package with Conda environments
#> 13                                                                  Example demonstrating how to create a package
#> 14                                                                     Example demonstrating how to use pipelines
#> 15                                                                  Example demonstrating how to use spatial data
#> 16                                                                       Example demonstrating conda environments
#> 17                                                                  Example demonstrating how to use spatial data
#> 18                                                                     Example demonstrating how to use timesteps
#> 19                                                                    Example demonstrating how to use iterations
#> 20                                                                                     LANDFIRE vegetation models
#> 21                                                              Runs the NetLogo agent-based modeling environment
#> 22                                                  Omni-directional habitat connectivity based on circuit theory
#> 23                                                   Connectivity planning for future climate and land-use change
#> 24                                                               The ST-Sim state-and-transition simulation model
#> 25                                                       Incorporates BurnP3+ fire simulations into ST-Sim models
#> 26 Integrates the Carbon Budget Model for the Canadian Forest Sector (CBM-CFS3) into the ST-Sim simulation model.
#> 27                                                                                Florida scrub jay habitat model
#> 28                         Integrates Potential Vegetation Fuels Change (PVFC) for LANDFIRE into the ST-Sim model
#> 29                                                   Integrates stocks and flows into the ST-Sim simulation model
#> 30                                                         Workbench for Integrated Species Distribution Modeling
#>    version
#> 1  Version
#> 2    0.0.0
#> 3    0.4.3
#> 4    0.4.4
#> 5    0.4.1
#> 6    1.1.1
#> 7   1.0.14
#> 8    2.3.2
#> 9    1.1.3
#> 10   1.0.2
#> 11   1.0.1
#> 12   1.0.0
#> 13   1.0.1
#> 14   1.0.0
#> 15   1.0.5
#> 16   1.0.0
#> 17   1.0.3
#> 18   1.0.0
#> 19   1.0.0
#> 20   1.1.2
#> 21   1.2.1
#> 22   1.0.1
#> 23  1.1.19
#> 24  3.3.11
#> 25   1.0.2
#> 26  1.0.17
#> 27   1.0.0
#> 28   1.0.2
#> 29  3.3.11
#> 30   1.0.5
#>                                                                          NA
#> 1                                                                  Location
#> 2                    C:\\Users\\birch\\SyncroSim\\Packages\\burnp3basemodel
#> 3                         C:\\Users\\birch\\SyncroSim\\Packages\\burnP3Plus
#> 4                C:\\Users\\birch\\SyncroSim\\Packages\\burnP3PlusCell2Fire
#> 5               C:\\Users\\birch\\SyncroSim\\Packages\\burnP3PlusPrometheus
#> 6                          C:\\Users\\birch\\SyncroSim\\Packages\\demosales
#> 7                           C:\\Users\\birch\\SyncroSim\\Packages\\designer
#> 8                              C:\\Users\\birch\\SyncroSim\\Packages\\dgsim
#> 9                                C:\\Users\\birch\\SyncroSim\\Packages\\epi
#> 10                C:\\Users\\birch\\SyncroSim\\Packages\\epiModelVocVaccine
#> 11                        C:\\Users\\birch\\SyncroSim\\Packages\\helloworld
#> 12                   C:\\Users\\birch\\SyncroSim\\Packages\\helloworldConda
#> 13                    C:\\Users\\birch\\SyncroSim\\Packages\\helloworldOmni
#> 14                C:\\Users\\birch\\SyncroSim\\Packages\\helloworldPipeline
#> 15                 C:\\Users\\birch\\SyncroSim\\Packages\\helloworldSpatial
#> 16            C:\\Users\\birch\\SyncroSim\\Packages\\helloworldSpatialConda
#> 17          C:\\Users\\birch\\SyncroSim\\Packages\\helloworldSpatialReports
#> 18                    C:\\Users\\birch\\SyncroSim\\Packages\\helloworldTime
#> 19             C:\\Users\\birch\\SyncroSim\\Packages\\helloworldUncertainty
#> 20                 C:\\Users\\birch\\SyncroSim\\Packages\\landfirevegmodels
#> 21                           C:\\Users\\birch\\SyncroSim\\Packages\\netlogo
#> 22                         C:\\Users\\birch\\SyncroSim\\Packages\\omniscape
#> 23                         C:\\Users\\birch\\SyncroSim\\Packages\\stconnect
#> 24                             C:\\Users\\birch\\SyncroSim\\Packages\\stsim
#> 25                   C:\\Users\\birch\\SyncroSim\\Packages\\stsimBurnP3Plus
#> 26                      C:\\Users\\birch\\SyncroSim\\Packages\\stsimcbmcfs3
#> 27 C:\\temp\\stsimFsj-bug\\ssimdebug\\.ssimdebug\\Packages\\stsimFsjHabitat
#> 28                         C:\\Users\\birch\\SyncroSim\\Packages\\stsimpvfc
#> 29                           C:\\Users\\birch\\SyncroSim\\Packages\\stsimsf
#> 30                             C:\\Users\\birch\\SyncroSim\\Packages\\wisdm
#>        NA
#> 1  Status
#> 2      OK
#> 3      OK
#> 4      OK
#> 5      OK
#> 6      OK
#> 7      OK
#> 8      OK
#> 9      OK
#> 10     OK
#> 11     OK
#> 12     OK
#> 13     OK
#> 14     OK
#> 15     OK
#> 16     OK
#> 17     OK
#> 18     OK
#> 19     OK
#> 20     OK
#> 21     OK
#> 22     OK
#> 23     OK
#> 24     OK
#> 25     OK
#> 26     OK
#> 27     OK
#> 28     OK
#> 29     OK
#> 30     OK

# List all the installed base packages
package(installed = "BASE")
#>                        name
#> 1                burnP3Plus
#> 2                 demosales
#> 3                  designer
#> 4                     dgsim
#> 5                       epi
#> 6                helloworld
#> 7           helloworldConda
#> 8            helloworldOmni
#> 9        helloworldPipeline
#> 10        helloworldSpatial
#> 11   helloworldSpatialConda
#> 12 helloworldSpatialReports
#> 13           helloworldTime
#> 14    helloworldUncertainty
#> 15        landfirevegmodels
#> 16                  netlogo
#> 17                omniscape
#> 18                stconnect
#> 19                    stsim
#> 20                    wisdm
#>                                                              description
#> 1                         Burn-P3+ package for burn probability modeling
#> 2                                      Demo Sales SyncroSim Base Package
#> 3                                             SyncroSim Package Designer
#> 4                         Simulates demographics of wildlife populations
#> 5                                Model of epidemic infections and deaths
#> 6                          Example demonstrating how to create a package
#> 7  Example demonstrating how to create a package with Conda environments
#> 8                          Example demonstrating how to create a package
#> 9                             Example demonstrating how to use pipelines
#> 10                         Example demonstrating how to use spatial data
#> 11                              Example demonstrating conda environments
#> 12                         Example demonstrating how to use spatial data
#> 13                            Example demonstrating how to use timesteps
#> 14                           Example demonstrating how to use iterations
#> 15                                            LANDFIRE vegetation models
#> 16                     Runs the NetLogo agent-based modeling environment
#> 17         Omni-directional habitat connectivity based on circuit theory
#> 18          Connectivity planning for future climate and land-use change
#> 19                      The ST-Sim state-and-transition simulation model
#> 20                Workbench for Integrated Species Distribution Modeling
#>    version
#> 1    0.4.3
#> 2    1.1.1
#> 3   1.0.14
#> 4    2.3.2
#> 5    1.1.3
#> 6    1.0.1
#> 7    1.0.0
#> 8    1.0.1
#> 9    1.0.0
#> 10   1.0.5
#> 11   1.0.0
#> 12   1.0.3
#> 13   1.0.0
#> 14   1.0.0
#> 15   1.1.2
#> 16   1.2.1
#> 17   1.0.1
#> 18  1.1.19
#> 19  3.3.11
#> 20   1.0.5

# List all available packages on the server (including currently installed)
package(installed = FALSE)
#>                     name
#> 1                   Name
#> 2             burnP3Plus
#> 3    burnP3PlusCell2Fire
#> 4   burnP3PlusPrometheus
#> 5              demosales
#> 6                  dgsim
#> 7                    epi
#> 8     epiModelVocVaccine
#> 9             helloworld
#> 10       helloworldConda
#> 11    helloworldPipeline
#> 12     helloworldSpatial
#> 13        helloworldTime
#> 14 helloworldUncertainty
#> 15     landfirevegmodels
#> 16                 lucas
#> 17               netlogo
#> 18             omniscape
#> 19             stconnect
#> 20                 stsim
#> 21       stsimBurnP3Plus
#> 22          stsimcbmcfs3
#> 23           stsimecodep
#> 24               stsimsf
#>                                                                                                                               description
#> 1                                                                                                                             Description
#> 2                                                                                          Burn-P3+ package for burn probability modeling
#> 3                                                                                  Cell2Fire fire growth model add-on package for BurnP3+
#> 4                                                                                 Prometheus fire growth model add-on package for BurnP3+
#> 5                                                                 Example SyncroSim Base Package of a very simple sales forecasting model
#> 6                                                                                          Simulates demographics of wildlife populations
#> 7                                                                                                 Model of epidemic infections and deaths
#> 8                                                                                      Integrates VOCs and vaccination into the Epi model
#> 9                                                                                           Example demonstrating how to create a package
#> 10                                                                                                       Example using conda environments
#> 11                                                                                                                Example using pipelines
#> 12                                                                                                             Example using spatial data
#> 13                                                                                                                Example using timesteps
#> 14                                                                                      Example using iterations to represent uncertainty
#> 15                                                                                                             LANDFIRE vegetation models
#> 16                                                                                         Land Use and Carbon Scenario Simulator (LUCAS)
#> 17                                                                                      Runs the NetLogo agent-based modeling environment
#> 18                                                                          Omni-directional habitat connectivity based on circuit theory
#> 19                                                                           Connectivity planning for future climate and land-use change
#> 20                                                                                       The ST-Sim state-and-transition simulation model
#> 21                                                       Example library simulating ST-Sim fire events with the BurnP3+ SyncroSim package
#> 22                          Integrates the Carbon Budget Model for the Canadian Forest Sector (CBM-CFS3) into the ST-Sim simulation model
#> 23 Calculates ecological departure in ST-Sim using the LANDFIRE Fire Regime Condition Class [*** REQUIRES SyncroSim v2.2.3 or higher ***]
#> 24                                                                           Integrates stocks and flows into the ST-Sim simulation model
#>    version
#> 1  Version
#> 2    0.4.3
#> 3    0.4.4
#> 4    0.4.4
#> 5    1.1.1
#> 6    2.3.2
#> 7    1.1.3
#> 8    1.0.2
#> 9    1.0.1
#> 10   1.0.0
#> 11   1.0.0
#> 12   1.0.5
#> 13   1.0.0
#> 14   1.0.0
#> 15   1.1.2
#> 16   1.0.6
#> 17   1.2.1
#> 18   1.0.1
#> 19  1.1.19
#> 20  3.3.11
#> 21   1.0.2
#> 22  1.0.17
#> 23   3.3.1
#> 24  3.3.11
 
# Check the package you're SsimLibrary is currently using
package(myLibrary)
#>    name                                      description version
#> 1 stsim The ST-Sim state-and-transition simulation model  3.3.11

# Check the templates available for an installed package
addPackage("helloworldSpatial")
#> Package <helloworldSpatial> is already installed
package(listTemplates = "helloworldSpatial")
#>                                name     displayName installed
#> 1 helloworldSpatial_example-library example-library       Yes
# }