Create or open a library

Creates or opens a SsimLibrary object. If summary = TRUE, returns library summary info. If summary = NULL, returns library summary info if ssimObject is a SsimLibrary, SsimLibrary object otherwise.

ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

# S4 method for SsimObject
ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

# S4 method for missingOrNULLOrChar
ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

Arguments

name
Character string, Project/Scenario/SsimLibrary. The path to a library or SsimObject.

summary
Logical. Default TRUE.

package
Character. The package type. The default is "stsim".

session
Session. If NULL, session() will be used.

addon
Character or character vector. One or more addons. See addon() for options.

forceUpdate
Logical. If FALSE (default) user will be prompted to approve any required updates. If TRUE, required updates will be applied silently.

overwrite
Logical. If TRUE an existing Library will be overwritten.

Value

Returns a SsimLibrary object.

Details

If name is SyncroSim Project or Scenario: Returns the SsimLibrary associated with the Project or Scenario.

If name is NULL: Create/open a SsimLibrary in the current working directory with the filename SsimLibrary.ssim.

If name is a string: If string is not a valid path treat as filename in working directory. If no file suffix provided in string then add .ssim. Attempts to open a library of that name. If library does not exist creates a library of type package in the current working directory.

If given a name and a package: Create/open a library called <name>.ssim. Returns an error if the library already exists but is a different type of package.

Examples

# \donttest{
# Create or open a library using the default session
myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"))

#> Warning: Default SyncroSim installation not found. Either install SyncroSim in the default location, or explicitly set the session path. See ?session for details.

#> Warning: SyncroSim not found.

# Create library using a specific session
mySession <- session()

#> Warning: Default SyncroSim installation not found. Either install SyncroSim in the default location, or explicitly set the session path. See ?session for details.

myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"), session = mySession)

#> Warning: SyncroSim not found.

session(myLibrary)

#> Warning: SyncroSim console could not be found at: SyncroSim not found./SyncroSim.Console.exe

#> [1] "SyncroSim not found."

filepath(myLibrary)

#> Warning: SyncroSim not found.

#> [1] "SyncroSim not found."

info(myLibrary)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("character"), new("standardGeneric", .Data = function (ssimLibrary) standardGeneric("info"), generic = structure("info", package = "rsyncrosim"),     package = "rsyncrosim", group = list(), valueClass = character(0),     signature = "ssimLibrary", default = NULL, skeleton = (function (ssimLibrary)     stop("invalid call in method dispatch to 'info' (no default method)",         domain = NA))(ssimLibrary)), <environment>): unable to find an inherited method for function ‘info’ for signature ‘"character"’

# }