Installed or available packages

Retrieves the packages installed or available for this version of SyncroSim.

package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for character
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for missingOrNULL
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for Session
package(ssimObject = NULL, installed = TRUE, listTemplates = NULL)

# S4 method for SsimLibrary
package(ssimObject)

Arguments

ssimObject
Session or SsimLibrary object. If NULL (default), session() will be used

installed
logical or character. TRUE (default) to list installed packages, FALSE to list available packages, and "BASE" to list installed base packages

listTemplates
character. Name of a SyncroSim package. If not NULL (default), then lists all templates available for that package. The package must be installed in the current Session. Ignored if ssimObject is a SsimLibrary object

Value

Returns a data.frame of packages installed or templates available for a specified package.

Examples

# \donttest{
# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# List all installed packages
package(mySession)

#>                       name
#> 1                      epi
#> 2                epiDataBc
#> 3            epiDataCanada
#> 4             epiDataWorld
#> 5          epiForecastIHME
#> 6    epiModelDeathHospital
#> 7       epiModelVocVaccine
#> 8               epiPublish
#> 9             epiTransform
#> 10              helloworld
#> 11      helloworldPipeline
#> 12    helloworldPipelinePy
#> 13       helloworldSpatial
#> 14     helloworldSpatialPy
#> 15          helloworldTime
#> 16        helloworldTimePy
#> 17   helloworldUncertainty
#> 18 helloworldUncertaintyPy
#> 19       landfirevegmodels
#> 20                   lucas
#> 21         modelKarlenPypm
#> 22                 netlogo
#> 23               stconnect
#> 24                   stsim
#> 25             stsimecodep
#> 26                 stsimsf
#>                                                                                 description
#> 1                                                                     Epidemic Base Package
#> 2                                                                          BC COVID-19 Data
#> 3                                                                    Canadian COVID-19 Data
#> 4                                                                       World COVID-19 Data
#> 5                                                                   IHME COVID-19 Forecasts
#> 6                                                             Death + Hospitalization Model
#> 7                                                                       VOC + Vaccine Model
#> 8                                                                          Data Publication
#> 9                                                                   Remove Seasonal Effects
#> 10                                            Example demonstrating how to create a package
#> 11                                               Example demonstrating how to use pipelines
#> 12                                   Example demonstrating how to use pipelines with python
#> 13                                            Example demonstrating how to use spatial data
#> 14                                            Example demonstrating how to use spatial data
#> 15                                               Example demonstrating how to use timesteps
#> 16                           Example demonstrating how to use timesteps with a python model
#> 17                                              Example demonstrating how to use iterations
#> 18                                  Example demonstrating how to use iterations with python
#> 19                                                               Landfire vegetation models
#> 20                                           Land Use and Carbon Scenario Simulator (LUCAS)
#> 21                                                                        Karlen PyPm Model
#> 22                                        Runs the NetLogo agent-based modeling environment
#> 23                             Connectivity planning for future climate and land-use change
#> 24                                         The ST-Sim state-and-transition simulation model
#> 25 Calculates ecological departure in ST-Sim using the LANDFIRE Fire Regime Condition Class
#> 26                             Integrates stocks and flows into the ST-Sim simulation model
#>    version
#> 1    1.1.0
#> 2    1.0.0
#> 3    1.0.0
#> 4    1.0.0
#> 5    1.0.0
#> 6    1.0.0
#> 7    1.0.0
#> 8    1.0.0
#> 9    1.0.0
#> 10   1.0.1
#> 11   1.0.0
#> 12   1.0.0
#> 13   1.0.1
#> 14   1.0.1
#> 15   1.0.0
#> 16   1.0.0
#> 17   1.0.0
#> 18   1.0.0
#> 19   1.1.0
#> 20   1.0.5
#> 21   1.0.0
#> 22   1.2.0
#> 23  1.1.14
#> 24   3.3.3
#> 25   3.3.0
#> 26   3.3.1

# List all the installed base packages
package(installed = "BASE")

#>                       name
#> 1                      epi
#> 2               helloworld
#> 3       helloworldPipeline
#> 4     helloworldPipelinePy
#> 5        helloworldSpatial
#> 6      helloworldSpatialPy
#> 7           helloworldTime
#> 8         helloworldTimePy
#> 9    helloworldUncertainty
#> 10 helloworldUncertaintyPy
#> 11       landfirevegmodels
#> 12                   lucas
#> 13                 netlogo
#> 14               stconnect
#> 15                   stsim
#>                                                       description version
#> 1                                           Epidemic Base Package   1.1.0
#> 2                   Example demonstrating how to create a package   1.0.1
#> 3                      Example demonstrating how to use pipelines   1.0.0
#> 4          Example demonstrating how to use pipelines with python   1.0.0
#> 5                   Example demonstrating how to use spatial data   1.0.1
#> 6                   Example demonstrating how to use spatial data   1.0.1
#> 7                      Example demonstrating how to use timesteps   1.0.0
#> 8  Example demonstrating how to use timesteps with a python model   1.0.0
#> 9                     Example demonstrating how to use iterations   1.0.0
#> 10        Example demonstrating how to use iterations with python   1.0.0
#> 11                                     Landfire vegetation models   1.1.0
#> 12                 Land Use and Carbon Scenario Simulator (LUCAS)   1.0.5
#> 13              Runs the NetLogo agent-based modeling environment   1.2.0
#> 14   Connectivity planning for future climate and land-use change  1.1.14
#> 15               The ST-Sim state-and-transition simulation model   3.3.3

# List all available packages on the server (including currently installed)
package(installed = FALSE)

#>                     name
#> 1              demosales
#> 2                  dgsim
#> 3             helloworld
#> 4     helloworldPipeline
#> 5      helloworldSpatial
#> 6         helloworldTime
#> 7  helloworldUncertainty
#> 8      landfirevegmodels
#> 9                  lucas
#> 10               netlogo
#> 11             stconnect
#> 12                 stsim
#> 13           stsimecodep
#> 14               stsimsf
#>                                                                                                                               description
#> 1                                                                 Example SyncroSim Base Package of a very simple sales forecasting model
#> 2                                                                                          Simulates demographics of wildlife populations
#> 3                                                                                           Example demonstrating how to create a package
#> 4                                                                                                                 Example using pipelines
#> 5                                                                                                              Example using spatial data
#> 6                                                                                                                 Example using timesteps
#> 7                                                                                       Example using iterations to represent uncertainty
#> 8                                                                                                                    Land Fire Veg Models
#> 9                                                                                          Land Use and Carbon Scenario Simulator (LUCAS)
#> 10                                                                                      Runs the NetLogo agent-based modeling environment
#> 11                                                                           Connectivity planning for future climate and land-use change
#> 12                                                                                       The ST-Sim state-and-transition simulation model
#> 13 Calculates ecological departure in ST-Sim using the LANDFIRE Fire Regime Condition Class [*** REQUIRES SyncroSim v2.2.3 or higher ***]
#> 14                                                                           Integrates stocks and flows into the ST-Sim simulation model
#>    version
#> 1    1.1.0
#> 2    2.3.0
#> 3    1.0.1
#> 4    1.0.0
#> 5    1.0.1
#> 6    1.0.0
#> 7    1.0.0
#> 8    1.1.0
#> 9    1.0.6
#> 10   1.2.0
#> 11  1.1.14
#> 12   3.3.3
#> 13   3.3.0
#> 14   3.3.1

 
# Check the package you're SsimLibrary is currently using
package(myLibrary)

#>    name                                      description version
#> 1 stsim The ST-Sim state-and-transition simulation model   3.3.3

# Check the templates available for an installed package
addPackage("helloworldSpatial")

#> Package <helloworldSpatial> is already installed

package(listTemplates = "helloworldSpatial")

#>              name     displayName installed
#> 1 example-library example-library       Yes

# }