Installed or available packages

Retrieves the packages installed or available in the current session if called on a Session object, or the packages added to a SyncroSim Library if called on a SsimLibrary object.

Usage

packages(ssimObject = NULL, installed = TRUE)

# S4 method for class 'character'
packages(ssimObject = NULL, installed = TRUE)

# S4 method for class 'missingOrNULL'
packages(ssimObject = NULL, installed = TRUE)

# S4 method for class 'Session'
packages(ssimObject = NULL, installed = TRUE)

# S4 method for class 'SsimLibrary'
packages(ssimObject)

Arguments

ssimObject

Session or SsimLibrary object. If NULL (default), session() will be used

installed

logical or character. TRUE (default) to list installed packages or FALSE to list available packages on the server

Value

Returns a data.frame of packages installed or templates available for a specified package.

Examples

# \donttest{
# Set the file path and name of the new SsimLibrary
myLibraryName <- file.path(tempdir(),"testlib")

# Set the SyncroSim Session and SsimLibrary
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)

# List all installed packages
packages(mySession)
#>                     name version
#> 1             burnP3Plus   2.0.0
#> 2    burnP3PlusCell2Fire   2.0.0
#> 3   burnP3PlusPrometheus   2.0.0
#> 4              demosales   2.0.0
#> 5     helloworldPipeline   2.0.0
#> 6      helloworldSpatial   2.0.0
#> 7         helloworldTime   2.0.0
#> 8  helloworldUncertainty   2.0.1
#> 9                  stsim   4.0.0
#> 10           stsimecodep   4.0.0
#> 11                 wisdm   2.0.0
#> 12                 wisdm   2.0.1
#> 13                 wisdm   2.0.3
#> 14                 wisdm   2.0.5
#> 15   wisdmStsimConnector   1.0.0
#>                                                                          description
#> 1                                                          Burn probability modeling
#> 2                                            Cell2Fire fire growth model for BurnP3+
#> 3                                           Prometheus fire growth model for BurnP3+
#> 4                                                       Demo Sales SyncroSim Package
#> 5                         Example demonstrating how to use pipelines with an R model
#> 6                      Example demonstrating how to use spatial data with an R model
#> 7                         Example demonstrating how to use timesteps with an R model
#> 8                        Example demonstrating how to use iterations with an R model
#> 9                                   The ST-Sim state-and-transition simulation model
#> 10 Calculates TNC's unified ecological departure from reference conditions in ST-Sim
#> 11                            Workbench for Integrated Species Distribution Modeling
#> 12                            Workbench for Integrated Species Distribution Modeling
#> 13                            Workbench for Integrated Species Distribution Modeling
#> 14                            Workbench for Integrated Species Distribution Modeling
#> 15                                        Connector package between WISDM and ST-Sim
#>                                                                                      location
#> 1             C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3Plus\\2.0.0
#> 2    C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3PlusCell2Fire\\2.0.0
#> 3   C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3PlusPrometheus\\2.0.0
#> 4              C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\demosales\\2.0.0
#> 5     C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldPipeline\\2.0.0
#> 6      C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldSpatial\\2.0.0
#> 7         C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldTime\\2.0.0
#> 8  C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldUncertainty\\2.0.1
#> 9                  C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\stsim\\4.0.0
#> 10           C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\stsimecodep\\4.0.0
#> 11                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.0
#> 12                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.1
#> 13                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.3
#> 14                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.5
#> 15                                                  C:\\gitprojects\\wisdmStsimConnector\\src
#>    status
#> 1      OK
#> 2      OK
#> 3      OK
#> 4      OK
#> 5      OK
#> 6      OK
#> 7      OK
#> 8      OK
#> 9      OK
#> 10     OK
#> 11     OK
#> 12     OK
#> 13     OK
#> 14     OK
#> 15     OK

# List all available packages on the server (including currently installed)
packages(installed = FALSE)
#>                     name version
#> 1             burnP3Plus   2.0.0
#> 2    burnP3PlusCell2Fire   2.0.0
#> 3   burnP3PlusPrometheus   2.0.0
#> 4              demosales   2.0.0
#> 5                  dgsim   3.0.0
#> 6             helloworld   2.0.0
#> 7             helloworld   2.0.1
#> 8     helloworldPipeline   2.0.0
#> 9      helloworldSpatial   2.0.0
#> 10        helloworldTime   2.0.0
#> 11      helloworldTimePy   2.0.0
#> 12 helloworldUncertainty   2.0.1
#> 13             omniscape   2.1.1
#> 14                 stsim   4.0.0
#> 15                 stsim   4.0.1
#> 16           stsimecodep   4.0.0
#> 17                 wisdm   2.0.1
#> 18                 wisdm   2.0.5
#> 19   wisdmStsimConnector   1.0.0
#>                                                                          description
#> 1                                     Burn-P3+ package for burn probability modeling
#> 2                                    Cell2Fire fire growth model package for BurnP3+
#> 3                                   Prometheus fire growth model package for BurnP3+
#> 4                                                       Demo Sales SyncroSim Package
#> 5                                     Simulates demographics of wildlife populations
#> 6                            Example demonstrating how to create a SyncroSim package
#> 7                            Example demonstrating how to create a SyncroSim package
#> 8                         Example demonstrating how to use pipelines with an R model
#> 9                      Example demonstrating how to use spatial data with an R model
#> 10                        Example demonstrating how to use timesteps with an R model
#> 11                    Example demonstrating how to use timesteps with a python model
#> 12                       Example demonstrating how to use iterations with an R model
#> 13                     Omni-directional habitat connectivity based on circuit theory
#> 14                                  The ST-Sim state-and-transition simulation model
#> 15                                  The ST-Sim state-and-transition simulation model
#> 16 Calculates TNC's unified ecological departure from reference conditions in ST-Sim
#> 17                            Workbench for Integrated Species Distribution Modeling
#> 18                            Workbench for Integrated Species Distribution Modeling
#> 19                                        Connector package between WISDM and ST-Sim
#>                                              url
#> 1           https://burnp3.github.io/BurnP3Plus/
#> 2           https://burnp3.github.io/BurnP3Plus/
#> 3           https://burnp3.github.io/BurnP3Plus/
#> 4           https://apexrms.github.io/demosales/
#> 5               https://apexrms.github.io/dgsim/
#> 6          https://apexrms.github.io/helloworld/
#> 7          https://apexrms.github.io/helloworld/
#> 8  https://apexrms.github.io/helloworldEnhanced/
#> 9  https://apexrms.github.io/helloworldEnhanced/
#> 10 https://apexrms.github.io/helloworldEnhanced/
#> 11 https://apexrms.github.io/helloworldEnhanced/
#> 12 https://apexrms.github.io/helloworldEnhanced/
#> 13          https://apexrms.github.io/omniscape/
#> 14                       https://docs.stsim.net/
#> 15                       https://docs.stsim.net/
#> 16                       https://docs.stsim.net/
#> 17              https://apexrms.github.io/wisdm/
#> 18              https://apexrms.github.io/wisdm/
#> 19              https://apexrms.github.io/wisdm/
 
# Check the package(s) in your SsimLibrary
packages(myLibrary)
#>    name                                      description version schema status
#> 1  core                           SyncroSim Core Package   3.0.9      3     OK
#> 2 stsim The ST-Sim state-and-transition simulation model   4.0.0      4     OK
# }