Create or return SyncroSim Session

Methods to create or return a SyncroSim Session.

Usage

session(x = NULL, silent = TRUE, printCmd = FALSE)

# S4 method for class 'missingOrNULLOrChar'
session(x = NULL, silent = TRUE, printCmd = FALSE)

# S4 method for class 'SsimObject'
session(x = NULL, silent = TRUE, printCmd = FALSE)

# S4 method for class 'Folder'
session(x = NULL, silent = TRUE, printCmd = FALSE)

session(ssimObject) <- value

# S4 method for class 'NULLOrChar'
session(ssimObject) <- value

# S4 method for class 'SsimObject'
session(ssimObject) <- value

Arguments

x

character or SsimObject. Path to SyncroSim installation. If NULL (default), then default path is used

silent

logical. Applies only if x is a path or NULL If TRUE, warnings from the console are ignored. Otherwise they are printed. Default is FALSE

printCmd

logical. Applies only if x is a path or NULL If TRUE, arguments passed to the SyncroSim console are also printed. Helpful for debugging. Default is FALSE

ssimObject

Project or Scenario object

value

Session object

Value

A SyncroSim Session object.

Details

In order to avoid problems with SyncroSim version compatibility and SsimLibrary updating, the new Session must have the same filepath as the Session of the SsimObject e.g. filepath(value)==filepath(session(ssimObject)). Therefore, the only time when you will need to set a new SyncroSim Session is if you have updated the SyncroSim software and want to update an existing SsimObject to use the new software.

Examples

# \donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session, SsimLibrary, and Project
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)
myProject <- project(myLibrary, project = "Definitions")

# Lists the folder location of SyncroSim Session
filepath(mySession)
#> [1] "C:/Program Files/SyncroSim Studio"

# Lists the version of SyncroSim Session
version(mySession)
#> [1] "3.0.9"

# Data frame of the packages installed with this version of SyncroSim
packages(mySession)
#>                     name version
#> 1             burnP3Plus   2.0.0
#> 2    burnP3PlusCell2Fire   2.0.0
#> 3   burnP3PlusPrometheus   2.0.0
#> 4              demosales   2.0.0
#> 5     helloworldPipeline   2.0.0
#> 6      helloworldSpatial   2.0.0
#> 7         helloworldTime   2.0.0
#> 8  helloworldUncertainty   2.0.1
#> 9                  stsim   4.0.0
#> 10                 stsim   4.0.1
#> 11           stsimecodep   4.0.0
#> 12                 wisdm   2.0.0
#> 13                 wisdm   2.0.1
#> 14                 wisdm   2.0.3
#> 15                 wisdm   2.0.5
#> 16   wisdmStsimConnector   1.0.0
#>                                                                          description
#> 1                                                          Burn probability modeling
#> 2                                            Cell2Fire fire growth model for BurnP3+
#> 3                                           Prometheus fire growth model for BurnP3+
#> 4                                                       Demo Sales SyncroSim Package
#> 5                         Example demonstrating how to use pipelines with an R model
#> 6                      Example demonstrating how to use spatial data with an R model
#> 7                         Example demonstrating how to use timesteps with an R model
#> 8                        Example demonstrating how to use iterations with an R model
#> 9                                   The ST-Sim state-and-transition simulation model
#> 10                                  The ST-Sim state-and-transition simulation model
#> 11 Calculates TNC's unified ecological departure from reference conditions in ST-Sim
#> 12                            Workbench for Integrated Species Distribution Modeling
#> 13                            Workbench for Integrated Species Distribution Modeling
#> 14                            Workbench for Integrated Species Distribution Modeling
#> 15                            Workbench for Integrated Species Distribution Modeling
#> 16                                        Connector package between WISDM and ST-Sim
#>                                                                                      location
#> 1             C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3Plus\\2.0.0
#> 2    C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3PlusCell2Fire\\2.0.0
#> 3   C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\burnP3PlusPrometheus\\2.0.0
#> 4              C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\demosales\\2.0.0
#> 5     C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldPipeline\\2.0.0
#> 6      C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldSpatial\\2.0.0
#> 7         C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldTime\\2.0.0
#> 8  C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\helloworldUncertainty\\2.0.1
#> 9                  C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\stsim\\4.0.0
#> 10                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\stsim\\4.0.1
#> 11           C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\stsimecodep\\4.0.0
#> 12                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.0
#> 13                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.1
#> 14                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.3
#> 15                 C:\\Users\\birch\\AppData\\Local\\SyncroSim Studio\\Packages\\wisdm\\2.0.5
#> 16                                                  C:\\gitprojects\\wisdmStsimConnector\\src
#>    status
#> 1      OK
#> 2      OK
#> 3      OK
#> 4      OK
#> 5      OK
#> 6      OK
#> 7      OK
#> 8      OK
#> 9      OK
#> 10     OK
#> 11     OK
#> 12     OK
#> 13     OK
#> 14     OK
#> 15     OK
#> 16     OK

# Set a new SyncroSim Session for the SyncroSim Project
session(myProject) <- session(x = filepath(session(myProject)))
# }