Create or return SyncroSim Session

Methods to create or return a SyncroSim Session.

session(x = NULL, silent = TRUE, printCmd = FALSE)

# S4 method for missingOrNULLOrChar
session(x = NULL, silent = TRUE, printCmd = FALSE)

# S4 method for SsimObject
session(x = NULL, silent = TRUE, printCmd = FALSE)

session(ssimObject) <- value

# S4 method for character
session(ssimObject) <- value

# S4 method for SsimObject
session(ssimObject) <- value

Arguments

x
character or SsimObject. Path to SyncroSim installation. If NULL (default), then default path is used

silent
logical. Applies only if x is a path or NULL If TRUE, warnings from the console are ignored. Otherwise they are printed. Default is FALSE

printCmd
logical. Applies only if x is a path or NULL If TRUE, arguments passed to the SyncroSim console are also printed. Helpful for debugging. Default is FALSE

ssimObject
Project or Scenario object

value
Session object

Value

A SyncroSim Session object.

Details

In order to avoid problems with SyncroSim version compatibility and SsimLibrary updating, the new Session must have the same filepath as the Session of the SsimObject e.g. filepath(value)==filepath(session(ssimObject)). Therefore, the only time when you will need to set a new SyncroSim Session is if you have updated the SyncroSim software and want to update an existing SsimObject to use the new software.

Examples

# \donttest{
# Specify file path and name of new SsimLibrary
myLibraryName <- file.path(tempdir(), "testlib")

# Set up a SyncroSim Session, SsimLibrary, and Project
mySession <- session()
myLibrary <- ssimLibrary(name = myLibraryName, session = mySession)
myProject <- project(myLibrary, project = "Definitions")

# Lists the folder location of SyncroSim Session
filepath(mySession)

#> [1] "C:/Program Files/SyncroSim"

# Lists the version of SyncroSim Session
version(mySession)

#> [1] "2.3.8"

# Data frame of the packages installed with this version of SyncroSim
package(mySession) 

#>                    name
#> 1       helloworldConda
#> 2    helloworldPipeline
#> 3     helloworldSpatial
#> 4        helloworldTime
#> 5 helloworldUncertainty
#> 6             stconnect
#> 7                 stsim
#> 8          stsimcbmcfs3
#> 9               stsimsf
#>                                                                                                      description
#> 1                                          Example demonstrating how to create a package with Conda environments
#> 2                                                                     Example demonstrating how to use pipelines
#> 3                                                                  Example demonstrating how to use spatial data
#> 4                                                                     Example demonstrating how to use timesteps
#> 5                                                                    Example demonstrating how to use iterations
#> 6                                                   Connectivity planning for future climate and land-use change
#> 7                                                               The ST-Sim state-and-transition simulation model
#> 8 Integrates the Carbon Budget Model for the Canadian Forest Sector (CBM-CFS3) into the ST-Sim simulation model.
#> 9                                                   Integrates stocks and flows into the ST-Sim simulation model
#>   version
#> 1   1.0.1
#> 2   1.0.0
#> 3   1.0.2
#> 4   1.0.0
#> 5   1.0.0
#> 6  1.1.16
#> 7   3.3.5
#> 8  1.0.12
#> 9   3.3.4

# Data frame of the base packages installed with this version of SyncroSim
package(mySession, installed = "BASE") 

#>                    name
#> 1       helloworldConda
#> 2    helloworldPipeline
#> 3     helloworldSpatial
#> 4        helloworldTime
#> 5 helloworldUncertainty
#> 6             stconnect
#> 7                 stsim
#>                                                             description version
#> 1 Example demonstrating how to create a package with Conda environments   1.0.1
#> 2                            Example demonstrating how to use pipelines   1.0.0
#> 3                         Example demonstrating how to use spatial data   1.0.2
#> 4                            Example demonstrating how to use timesteps   1.0.0
#> 5                           Example demonstrating how to use iterations   1.0.0
#> 6          Connectivity planning for future climate and land-use change  1.1.16
#> 7                      The ST-Sim state-and-transition simulation model   3.3.5

# Set a new SyncroSim Session for the SyncroSim Project
session(myProject) <- session(x = filepath(session(myProject)))
# }