Create or open a SsimLibrary

Creates or opens a SsimLibrary object. If summary = TRUE, returns SsimLibrary summary info. If summary = NULL, returns SsimLibrary summary info if ssimObject is a SsimLibrary, SsimLibrary object otherwise.

ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  template = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

# S4 method for SsimObject
ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  template = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

# S4 method for missingOrNULLOrChar
ssimLibrary(
  name = NULL,
  summary = NULL,
  package = NULL,
  session = NULL,
  addon = NULL,
  template = NULL,
  forceUpdate = FALSE,
  overwrite = FALSE
)

Arguments

name
SsimLibrary, Project or Scenario object, or character string (i.e. path to a SsimLibrary or SsimObject)

summary
logical. Default is TRUE

package
character. The package type. Default is "stsim"

session
Session object. If NULL (default), session() will be used

addon
character or character vector. One or more addon packages. See addon for options (optional)

template
character. Creates the SsimLibrary with the specified template (optional)

forceUpdate
logical. If FALSE (default) user will be prompted to approve any required updates. If TRUE, required updates will be applied silently

overwrite
logical. If TRUE an existing SsimLibrary will be overwritten

Value

Returns a SsimLibrary object.

Details

Example arguments:

If name is SyncroSim Project or Scenario: Returns the SsimLibrary associated with the Project or Scenario.

If name is NULL: Create/open a SsimLibrary in the current working directory with the filename SsimLibrary.ssim.

If name is a string: If string is not a valid path treat as filename in working directory. If no file suffix provided in string then add .ssim. Attempts to open a SsimLibrary of that name. If SsimLibrary does not exist creates a SsimLibrary of type package in the current working directory.

If given a name and a package: Create/open a SsimLibrary called <name>.ssim. Returns an error if the SsimLibrary already exists but is a different type of package.

Examples

# \donttest{
# Make sure packages are installed
addPackage("stsim")

#> Package <stsim> is already installed

addPackage("stsimsf")

#> Package <stsimsf> is already installed

# Create or open a SsimLibrary using the default Session
myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"))

# Create SsimLibrary using a specific Session
mySession <- session()

myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"),
                         session = mySession)

# Retrieve SsimLibrary properties
session(myLibrary)

#> class               : Session
#> filepath [character]: C:/Program Files/SyncroSim
#> silent [logical]    : TRUE
#> printCmd [logical]  : FALSE

filepath(myLibrary)

#> [1] "C:\\Users\\birch\\AppData\\Local\\Temp\\Rtmpo5Dja6/mylib.ssim"

info(myLibrary)

#>                    property
#> 1                     Name:
#> 2                    Owner:
#> 3            Last Modified:
#> 4                     Size:
#> 5                Read Only:
#> 6             Package Name:
#> 7      Package Description:
#> 8  Current Package Version:
#> 9  Minimum Package Version:
#> 10    External input files:
#> 11   External output files:
#> 12         Temporary files:
#> 13            Backup files:
#>                                                                    value
#> 1                                                                  mylib
#> 2                                                                    N/A
#> 3                                                  2021-09-09 at 4:53 PM
#> 4                                                    616 KB  (630,784 B)
#> 5                                                                     No
#> 6                                                                  stsim
#> 7                       The ST-Sim state-and-transition simulation model
#> 8                                                                  3.3.2
#> 9                                                                 3.2.18
#> 10  C:\\Users\\birch\\AppData\\Local\\Temp\\Rtmpo5Dja6\\mylib.ssim.input
#> 11 C:\\Users\\birch\\AppData\\Local\\Temp\\Rtmpo5Dja6\\mylib.ssim.output
#> 12   C:\\Users\\birch\\AppData\\Local\\Temp\\Rtmpo5Dja6\\mylib.ssim.temp
#> 13 C:\\Users\\birch\\AppData\\Local\\Temp\\Rtmpo5Dja6\\mylib.ssim.backup

# Load a SsimLibrary with addon package
myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"),
                         overwrite = TRUE, package = "stsim",
                         addon = "stsimsf")

# Create SsimLibrary from template
addPackage("helloworldEnhanced")

#> Package <helloworldEnhanced> is already installed

mySession <- session()
myLibrary <- ssimLibrary(name = file.path(tempdir(), "mylib"), 
                         session = mySession,
                         package = "helloworldEnhanced",
                         template = "example-library",
                         overwrite = TRUE)

#> Scenarios available in this template:    My Scenario

# }